requirements.txt
langchain
langchain-community
langchain-openai
langchain-qdrant
qdrant-client
python-dotenv
streamlit
pypdf
